Version 1.0.6
- fix for STM32 official package pin number range (int16_t)
- update of conditional compile conditions to macros of actual STM32 package

+++ src/GFX_ILI9486.cpp
@@ -23,7 +23,7 @@
 
 #define RGB_or_BGR MADCTL_BGR // BGR for my screen
 
-GFX_ILI9486::GFX_ILI9486(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin, int8_t bl_pin) :
+GFX_ILI9486::GFX_ILI9486(int16_t cs_pin, int16_t dc_pin, int16_t rst_pin, int16_t bl_pin) :
 #if defined(ESP8266)
   GFX_TFT_IO(320, 480, cs_pin, dc_pin, rst_pin)
 #else
@@ -34,14 +34,14 @@ GFX_ILI9486::GFX_ILI9486(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin, int8_t bl
   _bgr = RGB_or_BGR;
 }
 
-GFX_ILI9486::GFX_ILI9486(int8_t cs_pin, int8_t dc_pin, int8_t mosi_pin, int8_t sclk_pin, int8_t rst_pin, int8_t bl_pin) :
+GFX_ILI9486::GFX_ILI9486(int16_t cs_pin, int16_t dc_pin, int16_t mosi_pin, int16_t sclk_pin, int16_t rst_pin, int16_t bl_pin) :
   GFX_TFT_IO(320, 480, cs_pin, dc_pin, mosi_pin, sclk_pin, rst_pin, -1)
 {
   _bl_pin = bl_pin;
   _bgr = RGB_or_BGR;
 }
 
-GFX_ILI9486::GFX_ILI9486(SPIClass *spi, int8_t cs_pin, int8_t dc_pin, int8_t rst_pin, int8_t bl_pin) :
+GFX_ILI9486::GFX_ILI9486(SPIClass *spi, int16_t cs_pin, int16_t dc_pin, int16_t rst_pin, int16_t bl_pin) :
 #if defined(ESP8266)
   GFX_TFT_IO(320, 480, cs_pin, dc_pin, rst_pin)
 #else
