+++ src/GFX_ILI9806.h
@@ -16,9 +16,9 @@
 class GFX_ILI9806 : public GFX_TFT_IO
 {
   public:
-    GFX_ILI9806(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin = -1, int8_t bl_pin = -1);
-    GFX_ILI9806(int8_t cs_pin, int8_t dc_pin, int8_t mosi_pin, int8_t sclk_pin, int8_t rst_pin, int8_t bl_pin = -1);
-    GFX_ILI9806(SPIClass *spi, int8_t cs_pin, int8_t dc_pin, int8_t rst_pin = -1, int8_t bl_pin = -1);
+    GFX_ILI9806(int16_t cs_pin, int16_t dc_pin, int16_t rst_pin = -1, int16_t bl_pin = -1);
+    GFX_ILI9806(int16_t cs_pin, int16_t dc_pin, int16_t mosi_pin, int16_t sclk_pin, int16_t rst_pin, int16_t bl_pin = -1);
+    GFX_ILI9806(SPIClass *spi, int16_t cs_pin, int16_t dc_pin, int16_t rst_pin = -1, int16_t bl_pin = -1);
     GFX_ILI9806(GFX_TFT_IO_Plugin& plugin);
     void begin(uint32_t freq);
     void init(uint32_t freq = 0);
@@ -31,7 +31,7 @@ class GFX_ILI9806 : public GFX_TFT_IO
     void _writeCommandTransaction(uint8_t cmd);
     void _writeDataTransaction(uint8_t data);
   private:
-    int8_t _bl_pin;
+    int16_t _bl_pin;
     int8_t _bgr;
 };
