+++ src/GFX_OTM8009A.h
@@ -22,9 +22,9 @@
 class GFX_OTM8009A : public GFX_TFT_IO
 {
   public:
-    GFX_OTM8009A(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin = -1, int8_t bl_pin = -1);
-    GFX_OTM8009A(int8_t cs_pin, int8_t dc_pin, int8_t mosi_pin, int8_t sclk_pin, int8_t rst_pin, int8_t bl_pin = -1);
-    GFX_OTM8009A(SPIClass *spi, int8_t cs_pin, int8_t dc_pin, int8_t rst_pin = -1, int8_t bl_pin = -1);
+    GFX_OTM8009A(int16_t cs_pin, int16_t dc_pin, int16_t rst_pin = -1, int16_t bl_pin = -1);
+    GFX_OTM8009A(int16_t cs_pin, int16_t dc_pin, int16_t mosi_pin, int16_t sclk_pin, int16_t rst_pin, int16_t bl_pin = -1);
+    GFX_OTM8009A(SPIClass *spi, int16_t cs_pin, int16_t dc_pin, int16_t rst_pin = -1, int16_t bl_pin = -1);
     GFX_OTM8009A(GFX_TFT_IO_Plugin& plugin);
     void begin(uint32_t freq);
     void init(uint32_t freq = 0);
@@ -34,7 +34,7 @@ class GFX_OTM8009A : public GFX_TFT_IO
     void enableDisplay(bool enable);
     void setAddrWindow(uint16_t x, uint16_t y, uint16_t w, uint16_t h);
   private:
-    int8_t _bl_pin;
+    int16_t _bl_pin;
     int8_t _bgr;
 };
