+++ src/GFX_SPI_TFT/GFX_SPI_ILI9486.h
@@ -24,10 +24,10 @@
 class GFX_SPI_ILI9486 : public SPI_GFX_Class
 {
   public:
-    GFX_SPI_ILI9486(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin);
-    GFX_SPI_ILI9486(int8_t cs_pin, int8_t dc_pin, int8_t mosi_pin, int8_t sclk_pin, int8_t rst_pin = -1);
-    GFX_SPI_ILI9486(uint16_t width, uint16_t height, SPIClass *spi, int8_t cs_pin, int8_t dc_pin, int8_t rst_pin = -1);
-    GFX_SPI_ILI9486(uint16_t width, uint16_t height, int8_t cs_pin, int8_t dc_pin, int8_t mosi_pin, int8_t sclk_pin, int8_t rst_pin = -1);
+    GFX_SPI_ILI9486(int16_t cs_pin, int16_t dc_pin, int16_t rst_pin);
+    GFX_SPI_ILI9486(int16_t cs_pin, int16_t dc_pin, int16_t mosi_pin, int16_t sclk_pin, int16_t rst_pin = -1);
+    GFX_SPI_ILI9486(uint16_t width, uint16_t height, SPIClass *spi, int16_t cs_pin, int16_t dc_pin, int16_t rst_pin = -1);
+    GFX_SPI_ILI9486(uint16_t width, uint16_t height, int16_t cs_pin, int16_t dc_pin, int16_t mosi_pin, int16_t sclk_pin, int16_t rst_pin = -1);
     void setSpiKludge(bool rpi_spi16_mode = true); // call with false before init or begin to disable
     void begin(uint32_t freq);
     void init(uint32_t freq = 0);
