+++ src/GFX_ST7796S.cpp
@@ -24,7 +24,7 @@
 
 #define RGB_or_BGR MADCTL_BGR // BGR for my screen
 
-GFX_ST7796S::GFX_ST7796S(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin, int8_t bl_pin) :
+GFX_ST7796S::GFX_ST7796S(int16_t cs_pin, int16_t dc_pin, int16_t rst_pin, int16_t bl_pin) :
 #if defined(ESP8266)
   GFX_TFT_IO(320, 480, cs_pin, dc_pin, rst_pin)
 #else
@@ -35,14 +35,14 @@ GFX_ST7796S::GFX_ST7796S(int8_t cs_pin, int8_t dc_pin, int8_t rst_pin, int8_t bl
   _bgr = RGB_or_BGR;
 }
 
-GFX_ST7796S::GFX_ST7796S(int8_t cs_pin, int8_t dc_pin, int8_t mosi_pin, int8_t sclk_pin, int8_t rst_pin, int8_t bl_pin) :
+GFX_ST7796S::GFX_ST7796S(int16_t cs_pin, int16_t dc_pin, int16_t mosi_pin, int16_t sclk_pin, int16_t rst_pin, int16_t bl_pin) :
   GFX_TFT_IO(320, 480, cs_pin, dc_pin, mosi_pin, sclk_pin, rst_pin, -1)
 {
   _bl_pin = bl_pin;
   _bgr = RGB_or_BGR;
 }
 
-GFX_ST7796S::GFX_ST7796S(SPIClass *spi, int8_t cs_pin, int8_t dc_pin, int8_t rst_pin, int8_t bl_pin) :
+GFX_ST7796S::GFX_ST7796S(SPIClass *spi, int16_t cs_pin, int16_t dc_pin, int16_t rst_pin, int16_t bl_pin) :
 #if defined(ESP8266)
   GFX_TFT_IO(320, 480, cs_pin, dc_pin, rst_pin)
 #else
